Fix gosec int-to-uint32 overflow finding in moduloPartitioner
Convert the uint32 hash to int (always safe, a widening conversion)
before taking the modulo, instead of narrowing partitionCount (an
int) down to uint32 before the modulo. Same result, no possible
overflow, and satisfies golangci-lint's gosec G115 check without a
//nolint suppression.

+++ packages/scalability/partitioner.go
@@ -133,7 +133,7 @@ func NewModuloPartitioner(n int) (Partitioner, error) {
 
 // Partition implements Partitioner.
 func (p *moduloPartitioner) Partition(key string) int {
-	return int(hashString(key) % uint32(p.partitionCount))
+	return int(hashString(key)) % p.partitionCount
 }
 
 // PartitionCount implements Partitioner.
